requirements.txt
streamlit
python-docx
lxml
google-generativeai
python-dotenv